Making the initial commit, this includes files that are not used yet, but I will attempt to cut these out of the project before the final version.
Created new movement system and imported basic animations.
The biggest obstacle was in using the current scene, as the creator of the map did not set up colliders for the terrain and these would be a pain to set up unless done in Blender.
So I switched from mouse movement to keyboard movement.